[*] - divide widgets and provider from main.dart
